added plot_utility.py and updated modules:
-class_headers_to_dict.py: added function to return dictionary {title:
var} with items present in the read file. Used if plot_utility function
is called without x, y or misspelled variable labels.

-list_class_headers.py: added function that checks if the headers or two
files are the same.

-plot_utility.py: Implement function that plots the output of class
given a filename and two variable labels. It can compare with other
file.

-variable_title_dics.py: corrected the inverse dictionary: {var: title}

-wicmath.py: created module that store mathematical operations. At the
moment, it contains a function to obtain the relative_deviation between
two sets of data.